Next

+++ include/base/path_misc.h
@@ -30,25 +30,26 @@
 
   /* ... Include / Inclusion ........................................... */
 
-  #include "all_system.h"
-  #include "base/string_misc.h"
+     #include "all_system.h"
+     #include "base/string_misc.h"
+     #include <libgen.h>
 
 
   /* ... Functions / Funciones ......................................... */
 
-  int hash (const char *file, int nServ, int isfile);
+     int hash (const char *file, int nServ, int isfile);
 
-  int getFirstDir   ( char *dir, char *path);
-  int getSizeFactor ( char *name );
+     int  getFirstDir   ( char *dir, char *path);
+     long getSizeFactor ( char *name );
 
-  // get the last name of a path and erase the file name
-  int getNameFile(char *file, char *dir);
+     // get the last name of a path and erase the file name
+     int getNameFile(char *file, char *dir);
 
-  // get the first name of the path and erase the part name
-  int getNamePart(char *part, char *dir);
+     // get the first name of the path and erase the part name
+     int getNamePart(char *part, char *dir);
 
-  // erase the initial dir used in the url
-  int getDirWithURL(char *url, char *dir);
+     // erase the initial dir used in the url
+     int getDirWithURL(char *url, char *dir);
 
 
   /* .................................................................... */
@@ -58,3 +59,4 @@
   #endif
 
 #endif /* _PATH_MISC_H */
+
